Update
Fixed issue of players being able to sit on the same object. This was done by removing the seat tag from the object the player is currently targeting when they press the sit down key. Then adding the tag back when the button is released. One of the walls was rotated so the player can no longer walk through it and lock themselves out of the map. Additional scenes were added for the main menu. Buttons do not function currently.